+++ lib/utils/http_config_utils.dart
@@ -1,1 +1,0 @@
-export '../src/compatibility/http/http_config_utils.dart' show HttpConfigUtils;