+++ test/utils/dio/dio_advanced_features_test.dart
@@ -1,7 +1,7 @@
 import 'package:test/test.dart';
 import 'package:llm_dart_transport/dio.dart';
 import 'package:llm_dart/core/config.dart';
-import 'package:llm_dart/utils/http_config_utils.dart';
+import 'package:llm_dart/src/compatibility/http/http_config_utils.dart';
 import 'dio_proxy_test_stub.dart'
     if (dart.library.io) 'dio_proxy_test_io.dart'
     if (dart.library.html) 'dio_proxy_test_web.dart';