+++ test/utils/dio/dio_logging_test.dart
@@ -4,7 +4,7 @@ import 'package:llm_dart_transport/dio.dart';
 import 'package:llm_dart/core/config.dart';
 import 'package:llm_dart_transport/llm_dart_transport.dart'
     show Level, LogRecord, Logger, hierarchicalLoggingEnabled;
-import 'package:llm_dart/utils/http_config_utils.dart';
+import 'package:llm_dart/src/compatibility/http/http_config_utils.dart';
 import 'package:test/test.dart';
 
 void main() {