+++ test/utils/http_config_utils_test.dart
@@ -1,5 +1,5 @@
 import 'package:llm_dart/core/config.dart';
-import 'package:llm_dart/utils/http_config_utils.dart';
+import 'package:llm_dart/src/compatibility/http/http_config_utils.dart';
 import 'package:llm_dart_transport/dio.dart';
 import 'package:test/test.dart';